# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jasur Savdo API",
        "description": "API for Jasur Savdo project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/firms/{id}/change-debt": {
            "post": {
                "tags": [
                    "Firms"
                ],
                "summary": "Add a debt transaction for a firm",
                "description": "Add a new debt or payment transaction for a firm",
                "operationId": "ce4cca99d9cc06a7a6937430b676eaec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Firm ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "debt",
                                    "status"
                                ],
                                "properties": {
                                    "debt": {
                                        "type": "string",
                                        "example": "50000"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "oldi",
                                            "berdi"
                                        ],
                                        "example": "berdi"
                                    },
                                    "recorded_by": {
                                        "type": "string",
                                        "example": "Admin",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Debt transaction added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "firm_debt": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "firm_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "debt": {
                                                    "type": "string",
                                                    "example": "50000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "berdi"
                                                },
                                                "recorded_by": {
                                                    "type": "string",
                                                    "example": "Admin",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Firm not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Firm not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}